+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0 (2020.05.27)
+
+#### Features
+- Check internet connection before starting the update process
+- Implement clean and preview as separate commands
+
 # 0.1.0 (2020.05.26)
 
 #### Features